requirements.txt
# Python 3.11+ compatible - Updated Dec 2024
# TensorFlow 2.18 with CUDA 12.x support for GPU acceleration
tensorflow[and-cuda]==2.18.0
tensorflow-datasets==4.9.8
numpy==2.0.2
pandas==2.2.3
matplotlib==3.9.3
kaggle==1.6.17
tqdm==4.67.1
streamlit==1.41.1
pillow==11.0.0
scikit-learn==1.6.0
